switched test and main
